requirements.txt
flask
psutil
pywebview
requests
pyinstaller
